Rename file of views
