+++ client-angular/src/app/task-creation-wizard/step-2-label-assignment/label-assignment.component.ts
@@ -24,7 +24,7 @@ import {Router} from '@angular/router';
 import {LabelService} from 'app/labels/label.service';
 import {TaskCreationWizard} from 'app/task-creation-wizard/service/task-creation-wizard.service';
 import {WizardStep} from 'app/task-creation-wizard/wizard-step';
-import {LabelView} from 'proto/todolist/projections_pb';
+import {LabelView} from 'proto/todolist/views_pb';
 
 /**
  * A component which represents the second step of the Task Creation Wizard - a label assignment.
